+++ python/GafferUITest/__init__.py
@@ -134,6 +134,7 @@ from .BoxIOUITest import BoxIOUITest
 from .AnnotationsGadgetTest import AnnotationsGadgetTest
 from .PopupWindowTest import PopupWindowTest
 from .ColorChooserTest import ColorChooserTest
+from .ContextTrackerTest import ContextTrackerTest
 
 if __name__ == "__main__":
 	unittest.main()