+++ src/GafferModule/ContextProcessorBinding.cpp
@@ -76,6 +76,17 @@ ContextPtr nextIterationContextWrapper( Loop &loop )
 	return loop.nextIterationContext();
 }
 
+object previousIterationWrapper( Loop &loop, const Gaffer::ValuePlug &output )
+{
+	std::pair<const ValuePlug *, ContextPtr> result;
+	{
+		IECorePython::ScopedGILRelease gilRelease;
+		result = loop.previousIteration( &output );
+	}
+	return boost::python::make_tuple( ValuePlugPtr( const_cast<ValuePlug *>( result.first ) ), result.second );
+}
+
+
 ContextPtr inPlugContext( const ContextProcessor &n )
 {
 	IECorePython::ScopedGILRelease gilRelease;
@@ -196,6 +207,7 @@ void GafferModule::bindContextProcessor()
 	DependencyNodeClass<Loop>()
 		.def( "setup", &setupLoop )
 		.def( "nextIterationContext", &nextIterationContextWrapper )
+		.def( "previousIteration", &previousIterationWrapper )
 	;
 
 	DependencyNodeClass<ContextProcessor>()